# api/openapi.yaml
openapi: "3.0.3"
info:
  title: "InferGlow API"
  description: "InferGlow Agent Framework REST API"
  version: "6.0.0"
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /v1/memories/{id}:
    get:
      operationId: getMemory
      summary: Get a memory by ID
      tags: [Memories]
      parameters:
        - $ref: "#/components/parameters/MemoryId"
      responses:
        "200":
          description: Memory record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryRecord"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      operationId: deleteMemory
      summary: Delete a memory by ID
      tags: [Memories]
      parameters:
        - $ref: "#/components/parameters/MemoryId"
      responses:
        "200":
          description: Memory deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryDeleteResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /v1/flows:
    get:
      operationId: listFlows
      summary: List all registered flows
      tags: [Flows]
      responses:
        "200":
          description: Flow names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
    post:
      operationId: registerFlow
      summary: Register a new flow (hot-load YAML)
      tags: [Flows]
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema:
              type: string
              description: YAML flow definition
      responses:
        "201":
          description: Flow registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowRegisterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          description: Unprocessable flow definition
  /v1/flows/{name}:
    get:
      operationId: getFlow
      summary: Get a flow definition by name
      tags: [Flows]
      parameters:
        - $ref: "#/components/parameters/FlowName"
      responses:
        "200":
          description: Flow definition
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    MemoryId:
      name: id
      in: path
      required: true
      description: Memory ID
      schema:
        type: string
    FlowName:
      name: name
      in: path
      required: true
      description: Flow name
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServiceUnavailable:
      description: Service not configured
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required: [error]
...
    MemoryRecord:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        category:
          type: string
        facts:
          type: array
          items:
            type: string
...
    MemoryDeleteResponse:
      type: object
      properties:
        status:
          type: string
          example: "deleted"
        id:
          type: string
...
    FlowRegisterResponse:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          example: "registered"
...